debug

+++ pytorch_pretrained_bert/modeling.py
@@ -1222,7 +1222,7 @@ class BertForSequenceScoreTag(BertPreTrainedModel):
             all = torch.cat([i for i in all_aspect_berttag_list], 1)
             all= all.view(batch_size, num_aspect, max_seq_len, -1)
             print("+++++all.view(2, 3, 1, 2)+++++",all.size())
-            sequence_output = torch.mean(torch.FloatTensor(all), 1)
+            sequence_output = torch.mean(all, 1)
             print("+++++++sequence_output+++++",sequence_output.size())
             # tag_output = self.transformer(tag_output)
             # print("^^^^^^tag_transformer^^^^^^^", tag_output.size())
